Get rid of the check that all conduit responses start with a shield.
It looks like the conduit API changed recently(?) to remove the
'for(;;)' that used to live at the beginning of each response.  That's
ok.  We still check if the shielf is there, and remove it if so, but
if it's not there, we keep going.

Test plan:
run update_phabricator_repos.py and verify it works again.